* Avoid code duplication, * make byoc logic into plugin * align multi-source component registration with styleguide
