+++ packages/create-sitecore-jss/src/templates/nextjs-sxa/scripts/scaffold-component/plugins/byoc.ts
@@ -1,0 +1,19 @@
+import { ScaffoldComponentPlugin, ScaffoldComponentPluginConfig } from '..';
+import generateByocComponentSrc from '../../templates/byoc-component-src'; 
+
+/**
+ * Sets up generation for BYOC components
+ */
+class ByocPlugin implements ScaffoldComponentPlugin {
+  order = 0;
+
+  exec(config: ScaffoldComponentPluginConfig) {
+    if (config.args.includes('--byoc')) {
+      config.componentTemplateGenerator = generateByocComponentSrc;
+    }
+
+    return config;
+  }
+}
+
+export const byocPlugin = new ByocPlugin();
