+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/generate-component-builder/plugins/packages.ts
@@ -19,7 +19,10 @@ class PackagesPlugin implements ComponentBuilderPlugin {
      *      }
      *    ]
      *  }
-     * Or you can register all components from a path:
+     */
+
+    /**
+     * Or you can register all components from a path using the below approach:
      * import { PackageDefinition, getComponentList } from '@sitecore-jss/sitecore-jss-dev-tools/templating';
      * ...
      * const componentsPath = 'src/extra';
